main, parts: Include templaters in list of acceptable MIME types.
This fixes a bug where a MIME type not in the list of default templaters (SVG
in this case) would never be used, even if a custom templater was registered
for it.

Look up the available templaters at worker time, when formatting happens. This
ensures that any view-specific templaters are visible, and that nothing gets
left out of the decision to use a certain MIME type.